refacto(server): refactorize server detail ux/ui
