+++ src/features/servers/components/VirtualMachineCard.vue
@@ -1,0 +1,139 @@
+<script setup lang="ts">
+import { useI18n } from 'vue-i18n';
+import {
+  PlayIcon,
+  StopIcon,
+  ArrowPathIcon,
+  CubeIcon,
+} from '@heroicons/vue/24/outline';
+
+interface VirtualMachine {
+  id: string;
+  name: string;
+  state: string;
+  cpu: number;
+  memory: number;
+  storage: number;
+  ip: string;
+  os: string;
+}
+
+interface Props {
+  vm: VirtualMachine;
+}
+
+interface Emits {
+  (e: 'vm-action', vmId: string, action: 'start' | 'stop' | 'restart'): void;
+}
+
+defineProps<Props>();
+defineEmits<Emits>();
+
+const { t } = useI18n();
+
+const getStatusColor = (state: string) => {
+  switch (state) {
+    case 'running':
+      return 'text-emerald-600 bg-emerald-50 border-emerald-200';
+    case 'stopped':
+      return 'text-red-600 bg-red-50 border-red-200';
+    default:
+      return 'text-gray-600 bg-gray-50 border-gray-200';
+  }
+};
+
+const getMetricColor = (value: number) => {
+  if (value >= 80) return 'text-red-600 bg-red-100';
+  if (value >= 60) return 'text-amber-600 bg-amber-100';
+  return 'text-emerald-600 bg-emerald-100';
+};
+</script>
+
+<template>
+  <div
+    class="bg-white border border-slate-200 rounded-xl p-6 hover:shadow-md transition-shadow"
+  >
+    <div class="flex items-center justify-between mb-4">
+      <div class="flex items-center space-x-3">
+        <div class="p-2 bg-blue-100 rounded-lg">
+          <CubeIcon class="h-5 w-5 text-blue-600" />
+        </div>
+        <div>
+          <h4 class="text-lg font-semibold text-slate-900">
+            {{ vm.name }}
+          </h4>
+          <p class="text-sm text-slate-600">
+            {{ vm.os }} • {{ vm.ip }}
+          </p>
+        </div>
+      </div>
+
+      <div class="flex items-center space-x-3">
+        <span
+          :class="[
+            'px-3 py-1 text-xs font-semibold rounded-full border',
+            getStatusColor(vm.state),
+          ]"
+        >
+          {{ vm.state }}
+        </span>
+
+        <div class="flex space-x-1">
+          <button
+            v-if="vm.state === 'stopped'"
+            @click="$emit('vm-action', vm.id, 'start')"
+            class="p-2 text-emerald-600 hover:bg-emerald-50 rounded-lg transition-colors"
+            :title="t('servers.vm_start')"
+          >
+            <PlayIcon class="h-4 w-4" />
+          </button>
+          <button
+            v-if="vm.state === 'running'"
+            @click="$emit('vm-action', vm.id, 'stop')"
+            class="p-2 text-red-600 hover:bg-red-50 rounded-lg transition-colors"
+            :title="t('servers.vm_stop')"
+          >
+            <StopIcon class="h-4 w-4" />
+          </button>
+          <button
+            v-if="vm.state === 'running'"
+            @click="$emit('vm-action', vm.id, 'restart')"
+            class="p-2 text-blue-600 hover:bg-blue-50 rounded-lg transition-colors"
+            :title="t('servers.vm_restart')"
+          >
+            <ArrowPathIcon class="h-4 w-4" />
+          </button>
+        </div>
+      </div>
+    </div>
+
+    <div class="grid grid-cols-1 sm:grid-cols-4 gap-4">
+      <div class="text-center">
+        <p class="text-sm font-medium text-slate-600">CPU</p>
+        <p :class="['text-lg font-bold', getMetricColor(vm.cpu)]">
+          {{ vm.cpu }}%
+        </p>
+      </div>
+      <div class="text-center">
+        <p class="text-sm font-medium text-slate-600">{{ t('servers.memory') }}</p>
+        <p class="text-lg font-bold text-slate-900">
+          {{ (vm.memory / 1024).toFixed(1) }}GB
+        </p>
+      </div>
+      <div class="text-center">
+        <p class="text-sm font-medium text-slate-600">{{ t('servers.storage') }}</p>
+        <p class="text-lg font-bold text-slate-900">
+          {{ vm.storage }}GB
+        </p>
+      </div>
+      <div class="text-center">
+        <p class="text-sm font-medium text-slate-600">
+          {{ t('servers.ip_address') }}
+        </p>
+        <p class="text-lg font-bold text-slate-900">
+          {{ vm.ip }}
+        </p>
+      </div>
+    </div>
+  </div>
+</template>
